Set a network_mode of bridge for all services
Without setting a network_mode of bridge docker-compose starting with
version 2 of the api will start creating a network automatically based
on a scheme of [projectname]_default. This network appears to take an
incrementing IP address range (so the first one is 172.18.* after the
172.17.* taken by the bridge network). The network routing devtools
sets up doesn't work for ranges outside of 172.17.* so while containers
can talk to each other they can't be chatted with from the outside.

See also:
  https://github.com/docker/compose/issues/3012
  https://docs.docker.com/compose/compose-file/#network-mode
  https://docs.docker.com/compose/networking/